// Isa-win32-x64/resources/app/src/app/models/productModel.ts
export interface product{
    id?: number;
    name?: string;
    priceB?: string;
    priceS?: string;
    priceB$?: string;
    priceS$?: string;
    stock?: string;
    recibo?: string;
    create_at?: Date;
    description?: string;
    category?: string;
    img?: string;
    iva?: number;
    marca?: string;
    fecven?:"";
    marca2?: string;
    pricePro?: string;
}